FUSETOOLS2-2543 - refactor - extract classes to their own file
preparatory work for further work and introduce other tree items
