+++ src/extension/ExtensionContextHandler.ts
@@ -18,7 +18,8 @@ import * as KogitoVsCode from '@kie-tools-core/vscode-extension/dist';
 import { dirname } from 'path';
 import { execSync } from 'child_process';
 import { HelpFeedbackProvider } from '../views/providers/HelpFeedbackProvider';
-import { Integration, IntegrationsProvider } from '../views/providers/IntegrationsProvider';
+import { IntegrationsProvider } from '../views/providers/IntegrationsProvider';
+import { Integration } from '../views/providers/integrationTreeItems/Integration';
 import { NewCamelRouteCommand } from '../commands/NewCamelRouteCommand';
 import { NewCamelKameletCommand } from '../commands/NewCamelKameletCommand';
 import { NewCamelPipeCommand } from '../commands/NewCamelPipeCommand';
